requirements.txt
click==6.7
Flask==1.0.2
Flask-API==1.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
pkg-resources==0.0.0
PyJWT==1.6.4
Werkzeug==0.14.1
